moved GetTotalTokenizedDelegations
